rename user settings to app settings
