+++ Composer/packages/client/src/pages/setting/router.tsx
@@ -7,7 +7,7 @@ import { Router } from '@reach/router';
 import { ErrorBoundary } from '../../components/ErrorBoundary';
 
 import { DialogSettings } from './dialog-settings';
-import { UserSettings } from './user-settings';
+import { AppSettings } from './app-settings';
 import { RuntimeSettings } from './runtime-settings';
 
 const Routes = () => {
@@ -15,7 +15,7 @@ const Routes = () => {
     <ErrorBoundary>
       <Router>
         <DialogSettings path="dialog-settings" default />
-        <UserSettings path="preferences" />
+        <AppSettings path="preferences" />
         <RuntimeSettings path="runtime" />
       </Router>
     </ErrorBoundary>
